requirements.txt
google-generativeai>=0.8.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
streamlit>=1.40.0
requests>=2.32.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
